SearchBar: Add avatar to search bar
Alright, so technically the avatar code is already added to Settings but is missing
some parts that Google has kept inside Pixel Impl's. We don't want to introduce any
more reverse engineered code than we have to. Also, that heavily depends on the user
flashing GAPPS. This method works with or without gapps, every time :p

This commit kangs Google's method for showing the avatar as seen within the EmergencyInfo
repo.

https://github.com/DirtyUnicorns/android_packages_apps_EmergencyInfo/blob/q10x/src/com/android/emergency/preferences/EmergencyNamePreference.java#L213

- Refresh avatar imageview onResume
- Add click listener to link to User Settings for more options

Change-Id: Id87843ebce8c33e9670ce10cff38c614ed2c25e0
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>